test(ci): trim slow CPU L0 coverage
